refactor: dialog context 분리
