+++ src/atoms/dialog.ts
@@ -1,0 +1,4 @@
+import { DialogAtom } from "@/types/dialog";
+import { atom } from "jotai";
+
+export const dialogAtom = atom<DialogAtom | null>(null);
