+++ src/components/Dialog/index.ts
@@ -1,0 +1,3 @@
+export * from "./FeatureDialog";
+export * from "./GuideDialog";
+export * from "./HelpDialog";
