+++ src/types/dialog.ts
@@ -1,0 +1,6 @@
+export type DialogProps = {
+  open: boolean;
+  setOpen: (open: boolean) => void;
+};
+
+export type DialogAtom = "help" | "guide" | "feature";
